requirements.txt
pandas>=1.5.0
numpy>=1.24.0
requests>=2.28.0
python-telegram-bot>=20.0
beautifulsoup4>=4.11.0
selenium>=4.8.0
schedule>=1.2.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
scikit-learn>=1.0.0
scipy>=1.9.0
matplotlib>=3.5.0
seaborn>=0.11.0